Draft: move functions to draftgeoutils.linear_algebra
